Move caller identification-related code to the CallerIdentification folder
